"chore : refactored the types into different files for maintainability"
