+++ packages/plugin-injective/injective-sdk-client-ts/src/modules/explorer.ts
@@ -51,7 +51,7 @@ export async function getValidator(
     return this.request({
         method: (params: string) =>
             this.indexerGrpcExplorerApi.fetchValidator(params),
-        params: params.validatorAddress,
+        params: params.address,
     });
 }
