+++ packages/plugin-injective/injective-sdk-client-ts/src/types/base.ts
@@ -1,0 +1,20 @@
+import {
+    PaginationOption,
+
+} from "@injectivelabs/sdk-ts";
+export interface PaginationParams {
+    pagination?: PaginationOption;
+}
+export interface PaginationParams {
+    pagination?: PaginationOption;
+}
+export interface AddressParams {
+    address: string;
+}
+export interface TimeRangeParams {
+    startTime?: number;
+    endTime?: number;
+}
+export interface DenomParam {
+    denom: string;
+}
