+++ packages/plugin-injective/injective-sdk-client-ts/src/types/explorer.ts
@@ -1,25 +1,74 @@
 
+import {
+    Transaction,
+    ExplorerValidator,
+    ValidatorUptime,
+    PeggyDepositTx,
+    PeggyWithdrawalTx,
+    IBCTransferTx,
+    ExplorerStats,
+    ExchangePagination,
 
-
-//IBC params
-export interface GetDenomTraceParams {
+} from "@injectivelabs/sdk-ts";
+import { InjectiveExplorerRpc } from "@injectivelabs/indexer-proto-ts";
+import { TimeRangeParams } from "./base";
+/// Explorer Module Params
+export interface GetTxByHashParams {
     hash: string;
 }
 
-export interface GetDenomsTraceParams {
-    pagination?: PaginationOption;
+export interface GetAccountTxParams {
+    address: string;
+    limit?: number;
+    type?: string;
+    before?: number;
+    after?: number;
+    startTime?: number;
+    endTime?: number;
 }
 
-// Response interfaces
-export interface GetDenomTraceResponse {
-    denomTrace: DenomTrace;
+export interface GetValidatorUptimeParams {
+    validatorAddress: string;
 }
 
-export interface GetDenomsTraceResponse {
-    denomsTrace: DenomTrace[];
+export interface GetPeggyDepositTxsParams {
+    receiver?: string;
+    sender?: string;
+    limit?: number;
+    skip?: number;
 }
 
-export interface IBCTransferParams extends PaginationParams {
+export interface GetPeggyWithdrawalTxsParams {
+    sender?: string;
+    receiver?: string;
+    limit?: number;
+    skip?: number;
+}
+
+export interface GetBlocksParams {
+    before?: number;
+    after?: number;
+    limit?: number;
+    from?: number;
+    to?: number;
+}
+
+export interface GetBlockParams {
+    id: string;
+}
+
+export interface GetTxsParams {
+    before?: number;
+    after?: number;
+    limit?: number;
+    skip?: number;
+    type?: string;
+    startTime?: number;
+    endTime?: number;
+    chainModule?: string;
+}
+
+export interface GetIBCTransferTxsParams {
     sender?: string;
     receiver?: string;
     srcChannel?: string;
@@ -30,19 +79,71 @@ export interface IBCTransferParams extends PaginationParams {
     skip?: number;
 }
 
-export interface MsgIBCTransferParams {
-    amount: {
-        denom: string;
-        amount: string;
-    };
-    memo?: string;
-    sender: string;
-    port: string;
-    receiver: string;
-    channelId: string;
-    timeout?: number;
-    height?: {
-        revisionHeight: number;
-        revisionNumber: number;
-    };
+// Response interfaces
+export interface GetTxByHashResponse {
+    tx: Transaction;
+}
+
+export interface GetAccountTxResponse {
+    txs: Transaction[];
+    pagination: ExchangePagination;
+}
+
+export interface GetExplorerValidatorResponse {
+    validator: ExplorerValidator;
+}
+
+export interface GetValidatorUptimeResponse {
+    uptime: ValidatorUptime[];
+}
+
+export interface GetPeggyDepositTxsResponse {
+    txs: PeggyDepositTx[];
+}
+
+export interface GetPeggyWithdrawalTxsResponse {
+    txs: PeggyWithdrawalTx[];
+}
+
+export interface GetBlocksResponse
+    extends InjectiveExplorerRpc.GetBlocksResponse {}
+
+export interface GetBlockResponse
+    extends InjectiveExplorerRpc.GetBlockResponse {}
+
+export interface GetTxsResponse extends InjectiveExplorerRpc.GetTxsResponse {}
+
+export interface GetIBCTransferTxsResponse {
+    txs: IBCTransferTx[];
+}
+
+export interface GetExplorerStatsResponse {
+    stats: ExplorerStats;
+}
+
+export interface AccountTxParams extends TimeRangeParams {
+    address: string;
+    limit?: number;
+    type?: string;
+    before?: number;
+    after?: number;
+}
+
+export interface BlocksParams extends TimeRangeParams {
+    before?: number;
+    after?: number;
+    limit?: number;
+}
+
+export interface BlockParams {
+    id: string;
+}
+
+export interface TxsParams extends TimeRangeParams {
+    before?: number;
+    after?: number;
+    limit?: number;
+    skip?: number;
+    type?: string;
+    chainModule?: string;
 }
