+++ packages/plugin-injective/injective-sdk-client-ts/src/types/insurance.ts
@@ -1,4 +1,12 @@
 
+import {
+    InsuranceFund,
+    InsuranceModuleParams,
+} from "@injectivelabs/sdk-ts";
+import {
+    InjectiveOracleV1Beta1Oracle,
+} from "@injectivelabs/core-proto-ts";
+
 // Insurance Fund Module Params
 export interface GetInsuranceFundParams {
     marketId: string;
