+++ packages/plugin-injective/injective-sdk-client-ts/src/types/mint.ts
@@ -1,5 +1,6 @@
-
-
+import {
+    MinModuleParams,
+} from "@injectivelabs/sdk-ts";
 // Mint module params
 // Response interfaces
 export interface MintModuleParamsResponse {
@@ -12,62 +13,4 @@ export interface GetInflationResponse {
 
 export interface GetAnnualProvisionsResponse {
     annualProvisions: string;
-}
-
-// Token Factory Module Params// Param interfaces
-export interface GetDenomsFromCreatorParams {
-    creator: string;
-}
-
-export interface GetDenomAuthorityMetadataParams {
-    creator: string;
-    subDenom: string;
-}
-
-// Response interfaces
-export interface GetDenomsFromCreatorResponse {
-    denoms: string[];
-}
-
-export interface GetDenomAuthorityMetadataResponse {
-    metadata: AuthorityMetadata;
-}
-
-export interface GetTokenFactoryModuleParamsResponse {
-    params: TokenFactoryModuleParams;
-}
-
-export interface GetTokenFactoryModuleStateResponse {
-    state: TokenFactoryModuleState;
-}
-export interface TokenFactoryParams {
-    creator: string;
-    subDenom: string;
-}
-
-export interface MsgBurnParams {
-    amount: {
-        amount: string;
-        denom: string;
-    };
-}
-export interface MsgChangeAdminParams {
-    denom: string;
-    newAdmin: string;
-}
-
-export interface MsgCreateDenomParams {
-    subdenom: string;
-    decimals?: number;
-    name?: string;
-    symbol?: string;
-}
-export interface MsgMintParams {
-    totalAmount: {
-        amount: string;
-        denom: string;
-    };
-}
-export interface MsgSetDenomMetadataParams {
-    metadata: CosmosBankV1Beta1Bank.Metadata;
 }
